adds data-dialog-persistent-notification
